% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintbynumbr_scheme_generation.R
\name{color_2d_scale_hsv}
\alias{color_2d_scale_hsv}
\title{Title}
\usage{
color_2d_scale_hsv(
  color_x,
  color_y,
  grad_1_1 = 360,
  grad_1_2 = 180,
  grad_2_1 = 60,
  grad_2_2 = 240
)
}
\arguments{
\item{color_x}{Vector of first scale for color mixing; class numeric or factor; will convert factor or character values to numeric}

\item{color_y}{Vector of second scale for color mixing; class numeric or factor; will convert factor or character values to numeric. Must be equal length as coloy}

\item{grad_1_1}{"Starting point" of color for first gradient (corresponding to minimum value of color_x)}

\item{grad_1_2}{"Ending point" of color for first gradient (corresponding to maximum value of color_x)}

\item{grad_2_1}{"Starting point" of color for second gradient (corresponding to minimum value of color_y)}

\item{grad_2_2}{"Ending point" of color for second gradient (corresponding to minimum value of color_y)}
}
\value{

}
\description{
Generates a HSV-Scale Two-Dimensional Color Scale from Input Vectors
}
\examples{
data(iris)
iris_color_id <- color_2d_scale_hsv(color_x = iris$Sepal.Length,
   color_y = iris$Sepal.Width)
head(iris_color_id)

}
