% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintbynumbr_scheme_generation.R
\name{color_family_gen}
\alias{color_family_gen}
\title{Title}
\usage{
color_family_gen(level_1, level_2, start_angle = 0, stop_angle = 180)
}
\arguments{
\item{stop_angle}{}
}
\value{

}
\description{
Title
}
