% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintbynumbr_scheme_generation.R
\name{color_scale_2d_rgb}
\alias{color_scale_2d_rgb}
\title{color_scale_2d_rgb: Generate "2-Dimensional" Color Scales in RGB Color Space.}
\usage{
color_scale_2d_rgb(paint_by_1, paint_by_2)
}
\arguments{
\item{paint_by_1}{Vector (class numeric or factor-ideally ordered) containing values for the "X"
axis of the color gradient. Is a red-to-cyan gradient by default.}

\item{paint_by_2}{Vector (class numeric or factor--ideally ordered) containing values for the "Y"
axis of the color gradient. Is a yellow-to-blue gradient by default}
}
\value{
Returns a dataframe with the original input vectors, with an added column of hexidecimal
color codes ("color_id") to be plugged in to ggplot2 or other downstream functions.
(For now, also includes columns of the rescaled values used for the color gradients)
}
\description{
color_scale_2d_rgb: Generate "2-Dimensional" Color Scales in RGB Color Space.
}
\examples{
ex_paint_by_1  <- rnorm(n = 100)
ex_paint_by_2 <- rnorm(n = 100, mean = 5)
 color_scale_2d_rgb(ex_paint_by_1, ex_paint_by_2)
}
