% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintbynumbr_scheme_generation.R
\name{color_scale_2d_rgb_legend}
\alias{color_scale_2d_rgb_legend}
\title{color_scale_2d_rgb_legend: Legend for Generated "2-Dimensional" Color Scales in RGB Color Space.}
\usage{
color_scale_2d_rgb_legend(paint_by_1, paint_by_2, resolution = 20)
}
\arguments{
\item{paint_by_1}{Vector (class numeric or factor-ideally ordered) containing values for the "X"
axis of the color gradient. Is a red-to-cyan gradient by default.}

\item{paint_by_2}{Vector (class numeric or factor--ideally ordered) containing values for the "Y"
axis of the color gradient. It is a yellow-to-blue gradient by default}

\item{resolution}{Integer to determine number of entries on x and y axis of the color "legend" grid. Default is 20x20}
}
\value{
Returns a dataframe with a square grid of user-specified resolution mapped to the original input vectors,
with an added column of hexidecimal color codes ("color_id") to be plugged in to ggplot2
or other downstream functions using the "scale_<color/fill>_identity()" argument.
For now, x and y labels need to be provided manually during plotting.
}
\description{
color_scale_2d_rgb_legend: Legend for Generated "2-Dimensional" Color Scales in RGB Color Space.
}
\examples{
ex_paint_by_1  <- rnorm(n = 100)
ex_paint_by_2 <- rnorm(n = 100, mean = 5)
 color_scale_2d_rgb_legend(ex_paint_by_1, ex_paint_by_2)
}
