% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fars_accidents}
\alias{fars_accidents}
\title{Fatality Analysis Reporting System (FARS)}
\format{
A data frame with 2110 rows and 81 variables:
\describe{
\item{LONGITUD}{Longitude (deimal notation)}
\item{LATITUDE}{Latitude (decimal notation)}
...
}
}
\source{
\url{http://https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars/}
}
\usage{
fars_accidents
}
\description{
A dataset containing records of fatal automotive accidents in the US;
Unknown/unreported values removed
}
\keyword{datasets}
