% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintbynumbr_scheme_generation.R
\name{paint_by_2d_scale_hsv_legend}
\alias{paint_by_2d_scale_hsv_legend}
\title{Generates a Legend for HSV-Scale Two-Dimensional Color Scale from Input Vectors}
\usage{
paint_by_2d_scale_hsv_legend(
  paint_by_1,
  paint_by_2,
  grad_1_1 = 360,
  grad_1_2 = 180,
  grad_2_1 = 60,
  grad_2_2 = 240,
  resolution = 20
)
}
\arguments{
\item{paint_by_1}{Vector of first scale for color mixing; class numeric or factor; will convert factor or character values to numeric}

\item{paint_by_2}{Vector of second scale for color mixing; class numeric or factor; will convert factor or character values to numeric. Must be equal length as coloy}

\item{grad_1_1}{"Starting point" of color for first gradient (corresponding to minimum value of color_x)}

\item{grad_1_2}{"Ending point" of color for first gradient (corresponding to maximum value of color_x)}

\item{grad_2_1}{"Starting point" of color for second gradient (corresponding to minimum value of color_y)}

\item{grad_2_2}{"Ending point" of color for second gradient (corresponding to minimum value of color_y)}

\item{resolution}{Integer to determine number of entries on x and y axis of the color "legend" grid. Default is 20x20}
}
\value{
Data frame describing a square grid (with user-defined resolution --20x20 by default) with hexidecimal color identifies
showing the full "spectrum" of the color_scale_2d_rgb mapping function with specified inputs.
}
\description{
Generates a Legend for HSV-Scale Two-Dimensional Color Scale from Input Vectors
}
\examples{
data(iris)
iris_color_id <- paint_by_2d_scale_hsv(paint_by_1 = iris$Sepal.Length,
   paint_by_2= iris$Sepal.Width)
head(iris_color_id)

}
