% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintbynumbr_scheme_generation.R
\name{paintbynumbr_key}
\alias{paintbynumbr_key}
\title{Generates a Color Gradient "Key" from List of Hexidecimal Color IDs with Provided Input Values; To be Plotted Manually.}
\usage{
paintbynumbr_key(val_x, val_y)
}
\arguments{
\item{val_x}{Vector corresponding to input "X" scale used to generate \code{color_id} via e.g. \code{color_2d_scale_hsv}; same length as \code{color_id}.}

\item{val_y}{Vector corresponding to input "Y" scale used to generate \code{color_id} via e.g. \code{color_2d_scale_hsv}; same length as \code{color_id}.}

\item{color_id}{Vector of hexidecimal color IDs as characters (e.g. "#006DFF")}
}
\value{
3-Column data frame to be plotted (e.g. geom_tile) for color gradient with input values
}
\description{
Generates a Color Gradient "Key" from List of Hexidecimal Color IDs with Provided Input Values; To be Plotted Manually.
}
\examples{
paintbynumbr_key(1:200, 1:15) # works with uneven vector lengths
}
