% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintbynumbr_code.R
\name{remap}
\alias{remap}
\title{Remap}
\usage{
remap(
  input,
  input_start = min(input),
  input_end = max(input),
  output_start = 0,
  output_end = 1
)
}
\arguments{
\item{output_end}{}
}
\value{
A vector the length of the input vector that returns original
values scaled to be between \code{output_start} and \code{output_end}.
}
\description{
Remap
}
\examples{
example_vector <- rnorm(n = 10)
remap(input = example_vector, output_start = 0, output_end = 1)

}
