% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintbynumbr_scheme_generation.R
\name{test_scheme_RGB}
\alias{test_scheme_RGB}
\title{Title}
\usage{
test_scheme_RGB(color_1, color_2)
}
\arguments{
\item{color_1}{Vector (class numeric or factor-ideally ordered) containing values for the "X"
axis of the color gradient. Is a red-to-cyan gradient by default.}

\item{color_2}{Vector (class numeric or factor--ideally ordered) containing values for the "Y"
axis of the color gradient. Is a yellow-to-blue gradient by default}
}
\value{
Returns a dataframe with the original input vectors, with an added column of hexidecimal
color codes ("color_id") to be plugged in to ggplot2 or other downstream functions.
(For now, also includes columns of the rescaled values used for the color gradients)
}
\description{
Title
}
\examples{
ex_color_1  <- rnorm(n = 100)
ex_color_2 <- rnorm(n = 100, mean = 5)
 test_scheme_RGB(ex_color_1, ex_color_2)
}
