% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintbynumbr_scheme_generation.R
\name{color_family_gen}
\alias{color_family_gen}
\title{color_family_gen: Create Nested Color Schemes with Factors or Numbers}
\usage{
color_family_gen(
  level_1,
  level_2,
  start_angle = 0,
  stop_angle = 180,
  val_start = 0.4,
  sat_start = 0.5
)
}
\arguments{
\item{level_1}{Vector of numeric values, factors, or character (will be converted to factors) corresponding to "outer" layer of categorization (e.g. genus in genus-species pairs, or state in city-state pairs). Sets larger hue angle increments.}

\item{level_2}{Vector of numeric values, factors, or character (will be converted to factors) corresponding to "inner" layer of categorization (e.g. species in genus-species pairs, or city in a city-state relationship). Adjsuts color value and saturation, with small changes in hue. Must be of equal length to \code{level_1}}

\item{start_angle}{single numeric value (0-365); used to set color hue angle of minimum value in \code{level_1} input. Must be smaller than \code{stop_angle.}}

\item{stop_angle}{Single numeric value (0-365); used to set color hue angle of maximum value in \code{level_1} input. Must be greater than \code{start_angle.}}

\item{val_start}{Single numeric value (0-1.0): used to set value of darkest point; lower values are closer to black}

\item{sat_start}{Single numeric value (0-1.0): used to set value of least saturated point; lower values are closer to grey}
}
\value{
Returns a vector of hexidecimal color codes of equal length of input vectors.
}
\description{
color_family_gen: Create Nested Color Schemes with Factors or Numbers
}
\examples{
ex_level_1 <- rep(c("A","B","C","D"), 25)
ex_level_2 <- rnorm(n = 100)
}
